protondx/requirements.txt
alabaster==0.7.12
appnope==0.1.0
asgiref==3.2.7
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bleach==3.1.5
botocore==1.15.49
cachetools==4.1.0
cement==2.8.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.5
django-dotenv==1.4.2
django-filter==2.2.0
django-filters==0.2.1
django-import-export==2.0.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
future==0.16.0
google-auth==1.4.2
google-colab==1.0.0
gunicorn==20.0.4
idna==2.7
imagesize==1.2.0
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.6
notebook==5.2.2
numpy==1.18.3
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==20.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
portpicker==1.2.0
postcodes-io-api==0.0.4
prometheus-client==0.7.1
prompt-toolkit==1.0.18
psycopg2==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.0
python-dotenv==0.10.3
pythononwheels==0.922
pytz==2020.1
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.20.1
rsa==4.0
scipy==1.4.1
seaborn==0.10.1
semantic-version==2.5.0
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.17.0
six==1.11.0
snowballstemmer==2.0.0
Sphinx==3.0.4
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
tablib==1.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==4.5.3
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.24.3
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
django-widget-tweaks==1.4.8

